+++ samples/databus/blob-storage-databus-cleanup-function/ABSDataBus_7/SenderAndReceiver/MessageWithLargePayload.cs
@@ -3,8 +3,6 @@
 [TimeToBeReceived("00:00:10")]
 public class MessageWithLargePayload : ICommand
 {
-#pragma warning disable CS0618 // Type or member is obsolete
-    public DataBusProperty<byte[]> LargePayload { get; set; }
-#pragma warning restore CS0618 // Type or member is obsolete
+    public ClaimCheckProperty<byte[]> LargePayload { get; set; }
     public string Description { get; set; }
 }
